fix: remove 403 status code from default config
As suggested in [#1157](https://github.com/lycheeverse/lychee/issues/1157) and merged in [#1377](https://github.com/lycheeverse/lychee/pull/1377), the HTTP status code 403 was removed from the default
configuration.

With this change, the default configuration in the docs will follow what
is described at Commandline Parameters in the README of lycheeverse/lychee.