format markdown as rich text, rather than code. i think,
that this is generally nicer to read. However, this means that the
website text is rendered from the same text as the CLI --help. The
website understands Markdown, but the CLI does not (obviously) so we
have to be careful.

I would suggest that when writing help, the focus should be on
readability within the *console*. This means that while Markdown can be
used, it should be limited to syntax which is unobtrusive in raw text.
Raw text is still the main format which the help text will be  rendered
in.

Personally, these would be okay:
- single `*` or `_` for emphasis (with preference for `*`)
- single backticks for inline code
- four space indentation for code blocks
- bullet and numbered lists

Imo, these would *not* be okay, because they appear too jarring
in plain text:
- code blocks with triple backtick fences. this includes any
  astro-specific asides and the like.
- link syntax with `[link](https://url)`
- bold when used for subheadings like `**Note**:`

I think this is a good compromise which lets the same text be usable for
both CLI --help and the website's rich text HTML.