Fetch readme and generate Markdown for CLI flags at build time (#113)
* use custom integration

* add message

* rm cli.md

* trying single code block highlight again. why is it blue???

* steal starlight's colours to make it grey ;-;

* Revert "steal starlight's colours to make it grey ;-;"

This reverts commit 2b85c81f86ebae7bd12c39247fa3afdf8adcd092.

* Revert "trying single code block highlight again. why is it blue???"

This reverts commit 35353c7392d2a53dbd646b8ec1cfd9ac1b6f0c9e.

* biome

* remove escapeMarkdown

* remove default value formatting

* try add remark-smartypants to package.json

* Revert "remove default value formatting"

This reverts commit 0936b318a95b21226e65bb2edf16524d1466fa50.

* format markdown as rich text, rather than code. i think,

that this is generally nicer to read. However, this means that the
website text is rendered from the same text as the CLI --help. The
website understands Markdown, but the CLI does not (obviously) so we
have to be careful.

I would suggest that when writing help, the focus should be on
readability within the *console*. This means that while Markdown can be
used, it should be limited to syntax which is unobtrusive in raw text.
Raw text is still the main format which the help text will be  rendered
in.

Personally, these would be okay:
- single `*` or `_` for emphasis (with preference for `*`)
- single backticks for inline code
- four space indentation for code blocks
- bullet and numbered lists

Imo, these would *not* be okay, because they appear too jarring
in plain text:
- code blocks with triple backtick fences. this includes any
  astro-specific asides and the like.
- link syntax with `[link](https://url)`
- bold when used for subheadings like `**Note**:`

I think this is a good compromise which lets the same text be usable for
both CLI --help and the website's rich text HTML.

* pnpm lock

* reorder sidebar based on (subjective) frequency of use. also

tweak page titles to be in Title Case.

* extract lychee-version.ts and add note of current version to docs

* Apply suggestions from code review

Co-authored-by: Matthias Endler <matthias@endler.dev>

* undo 3 blank lines

---------

Co-authored-by: Matthias Endler <matthias@endler.dev>